[Refactoring & Docs] 전체적인 파일 수정 및 코드 오류 수정
