+++ Account.cpp
@@ -6,14 +6,12 @@ using namespace std;
 
 // 김현수
 Account::Account(const Account &ref)
-    : accountID(ref.accountID), balance(ref.balance)
-{                                                        //복사 생성자
+    : accountID(ref.accountID), balance(ref.balance) {   //복사 생성자
   customerName = new char[strlen(ref.customerName) + 1]; //널문자 자리 하나
   strcpy(customerName, ref.customerName);                // 깊은 복사
 }
 Account::Account(int ID, int money, char *name)
-    : accountID(ID), balance(money)
-{                                            // 생성자
+    : accountID(ID), balance(money) {        // 생성자
   customerName = new char[strlen(name) + 1]; //널문자 자리 하나
   strcpy(customerName, name);                // 깊은 복사
 }
@@ -24,14 +22,11 @@ int Account::GetAccID() const { return this->accountID; }
 void Account::Deposit(int money) { balance += money; }
 
 // 김지은
-int Account::Withdraw(int money)
-{
-  if (balance < money)
-  { //잔액<출금액
+int Account::Withdraw(int money) {
+  if (balance < money) { //잔액<출금액
     return -1;
   }
-  if (balance == money)
-  {
+  if (balance == money) {
     return 0; //잔액=출금액 일시 0 반환
   }
   balance -= money;
@@ -39,27 +34,8 @@ int Account::Withdraw(int money)
 }
 
 // 유진
-void Account::ShowAccInfo() const
-{
+void Account::ShowAccInfo() const {
   cout << "계좌 : " << accountID << endl;
   cout << "잔액 : " << balance << endl;
   cout << "이름 : " << customerName << endl;
-}
-
-//최지우
-void AccountHandler::ShowAllAccInfo() const
-{
-  for (int i = 0; i < accNum; i++)
-  {
-    accArr[i]->ShowAccInfo();
-    cout << endl;
-  }
-}
-
-AccountHandler::~AccountHandler()
-{
-  for (int i = 0; i < accNum; i++)
-  {
-    delete accArr[i];
-  }
 }
