+++ Account.h
@@ -20,19 +20,4 @@ public:
   ~Account();
 };
 
-class AccountHandler {
-private:
-  Account *accArr[100];
-  int accNum;
-
-public:
-  AccountHandler();
-  void ShowMenu(void) const;
-  void MakeAccount(void);
-  void DepositMoney(void);
-  void WithdrawMoney(void);
-  void ShowAllAccInfo(void) const;
-  ~AccountHandler();
-};
-
 #endif
