+++ AccountHandler.h
@@ -1,0 +1,21 @@
+#ifndef __ACCOUNTHANDLER_H__
+#define __ACCOUNTHANDLER_H__
+
+#include "Account.h"
+
+class AccountHandler {
+private:
+  Account *accArr[100];
+  int accNum;
+
+public:
+  AccountHandler();
+  void ShowMenu(void) const;
+  void MakeAccount(void);
+  void DepositMoney(void);
+  void WithdrawMoney(void);
+  void ShowAllAccInfo(void) const;
+  ~AccountHandler();
+};
+
+#endif
