+++ main.cpp
@@ -1,4 +1,4 @@
-#include "Account.h"
+#include "AccountHandler.h"
 #include <iostream>
 
 using namespace std;
